fix #2 - use AbstractCRUDTestController from core
